beginnings of addr_expr
It is a partial copy from tree-object-size and many things are broken
because of the constant assumption in the source.  It does however
work well with passthrough calls and with PHIs, although the
functionality is limited.  Getting inner references is still broken,
that's the next thing I'm going to work on fixing.

Also, ADDR_EXPR starts bringing in some object size related arithmetic
and the way I'm going to deal with it is to have saturated arithmetic.
Basically, all subtractions will be (obj - MIN(obj, off)) so that
there is no overflow.  Similar saturated trick would be needed for
addition and multiplication.  This ought to take care of undefined
behaviour when computing object bounds.  With those expressions in
place, using ranges ought to be safe too, I think...

If this works, we could drop tree-object-size altogether in the end.