tree-dynamic-object-size: Support ADDR_EXPR
It is a partial copy from tree-object-size and supports computation of
sizes of whole objects.

One key difference from constant-size ADDR_EXPR computation is the
computation of residual sizes based on offset.  For bdos we attempt to
compute an expression and to try and ensure that it does not overflow,
saturated arithmetic is used.