tree-dynamic-object-size: Handle GIMPLE_PHI
Where the target object is a PHI node, set the size to be a PHI node
with sizes of the target object.  All object size information is now
split into an SSA name and its size expression.  This allows non-SSA
size expressions to be evaluated all at once in the end and generated
statements inserted into the right places before returning the SSA
tree representing the size.

This change also adds support for dependency loops since that is a
common occurrence with PHI nodes.  Dependency loop handling is single
pass unlike tree-object-size due to the separation of size SSA names
and actual expressions, which allow for placeholders and unresolved
dependency loop counting in case of self-reference.  The downside
though is that if there's an unresolved dependency loop and the object
size computation fails, all of the computations are scrapped.  Ideally
only computations involved in the unresolved dependency loop should be
scrapped but I don't know if it's worth the extra state that would
have to be managed.  We could revisit this if it ends up being a
bottleneck.

gcc/ChangeLog:

	* tree-dynamic-object-size.c: Include gimplify-me.h.
	(object_size_info): New member deploop.
	(object_size_exprs): New vector of object size expressions.
	(emit_size_stmts, emit_size_phi_stmts, eval_size_expr,
	gimplify_size_exprs): New functions.
	(compute_builtin_dyn_object_size): New argument FUN.
	Grow object_size_exprs and call gimplify_size_exprs.
	(maybe_update_dependency_loop, cache_object_size,
	cache_object_size_copy): New functions.
	(set_object_size_ssa, call_object_size): Call
	cache_object_size_copy.
	(collect_object_sizes_for): Handle GIMPLE_PHI.
	(init_dynamic_object_sizes): Allocate object_size_exprs.
	(fini_dynamic_object_sizes): Deallocate object_size_exprs.
	* gcc/tree-dynamic-object-size.h
	(compute_builtin_dyn_object_size): New argument.

gcc/testsuite/ChangeLog:

	* gcc.dg/builtin-dynamic-object-size-0.c
	(test_builtin_malloc_condphi, test_builtin_malloc_condphi2,
	test_builtin_malloc_condphi3, test_builtin_calloc_condphi,
	test_deploop): New tests.
	(main): Call them.

Signed-off-by: Siddhesh Poyarekar <siddhesh@gotplt.org>