tree-dynamic-object-size: Support ADDR_EXPR
This is the beginnings of support for ADDR_EXPR objects and is partly
based on the implementation in tree-object-size, splitting out
functions for readability.

One key difference from constant-size ADDR_EXPR computation is the
computation of residual sizes based on offset.  This pass attempts to
compute an expression with guards to try and not overflow.  This is
however still rudimentary and a subsequent patch for subobject support
makes it more comprehensive by handling negative offsets as well.

The size expressions based on offsets may look arbitrarily complex but
in practice most parts of the expression tend to fold away due to
being constants.  Despite that it is still a potential bottleneck and
may need some artifical backstop (such as bailing out on computation
if the size expression nests more than an arbitrarily chosen N levels)
to reduce compile time as well as avoid adding too much of a
performance overhead to generated code.

gcc/ChangeLog:

	* builtins.c (fold_builtin_dyn_object_size): Handle ADDR_EXPR.
	* tree-dynamic-object-size.c (compute_object_offset,
	size_for_offset, get_whole_var, whole_var_size,
	addr_dyn_object_size): New functions.
	(compute_builtin_dyn_object_size): Handle ADDR_EXPR.
	(expr_object_size): New function.
	(collect_object_sizes_for): Use it.

gcc/testsuite/ChangeLog:

	* gcc.dg/builtin-dynamic-object-size-0.c
	(test_builtin_malloc_condphi4, test_builtin_malloc_condphi5,
	test_passthrough_nonssa): New tests.
	(main): Call them.

Signed-off-by: Siddhesh Poyarekar <siddhesh@gotplt.org>