tree-dynamic-object-size: Handle GIMPLE_ASSIGN
Follow assignment operations and pointer-plus operations to get the
effective object size.  Unlike tree-object-size, this pass handles
negative offsets correctly and returns a size within the bounds of the
parent object; offsets > SIZE_MAX / 2 are considered negative to
emulate ssize_t.

To make negative offsets work, all object size computations store the
"whole variable" size to ensure that the negative offsets don't go
beyond bounds of the object within which PTR points.  As mentioned in
the previous patch, this makes the size_for_offset expressions quite
complex, but the aim at the moment is to be more correct than fast.
We can tweak this later if it is found to have a noticeable
performance impact.

When gcc is invoked without optimization, attempt to minimally
recognize GIMPLE_ASSIGN and ADDR_EXPR expressions so that
__builtin_dynamic_object_size is at least minimally useful without
optimization.

gcc/ChangeLog:

	* tree-dynamic-object-size.c (object_whole_sizes,
	object_whole_size_exprs): New vectors.
	(size_for_offset): Support negative offsets.
	(whole_var_size): Adjust.
	(addr_dyn_object_size): New argument wholesizep.
	(emit_size_stmts): New arguments wholesize_ssa and
	wholesize_expr.  Emit statements for whole size.
	(emit_size_phi_stmts): Likewise.
	(size_bound_expr): New function.
	(eval_size_expr): Call it.  New arguments wholesize and
	wholesize_expr.
	(gimplify_size_exprs): Adjust for whole sizes.
	(compute_builtin_dyn_object_size): Do simple object size
	computation for unoptimized case.
	(expr_object_size, cache_object_size,cache_object_size_copy,
	set_object_size_ssa, call_object_size): Adjust for wholesize.
	(make_object_size_name, plus_stmt_object_size): New functions.
	(collect_object_sizes_for): Support GIMPLE_ASSIGN.  Adjust for
	wholesize.
	(init_dynamic_object_sizes): Allocate object_wholesizes and
	object_wholesize_exprs.
	(fini_dynamic_object_sizes): Deallocate object_wholesizes and
	object_wholesize_exprs.

gcc/testsuite/ChangeLog

	* gcc.dg/builtin-dynamic-object-size-0.c (dynarray_struct):
	New struct.
	(test_dynarray_struct, test_substring,
	test_substring_ptrplus): New tests.
	(main): Call them.

Signed-off-by: Siddhesh Poyarekar <siddhesh@gotplt.org>