compute_builtin_object_size: Return a tree
Transform tree-object-size to operate on tree objects instead of host
wide integers.  This makes it easier to extend to dynamic expressions
for object sizes.

The compute_builtin_object_size interface also now returns a tree
expression instead of HOST_WIDE_INT, so callers have been adjusted to
account for that.

Signed-off-by: Siddhesh Poyarekar <siddhesh@gotplt.org>