tree-object-size: More consistent behaviour with flex arrays
The tree object size pass tries to fail when it detects a flex array in
the struct, but it ends up doing the right thing only when the flex
array is in the outermost struct.  For nested cases (such as arrays
nested in a union or an inner struct), it ends up taking whatever value
the flex array is declared with, using zero for the standard flex array,
i.e. [].

Change this behaviour to make it more consistent across the board,
honoring -fstrict-flex-arrays.  With this change, any array at the end
of the struct will end up causing __bos to use the allocated value of
the outer object, bailing out in the maximum case when it can't find it.
In the minimum case, it will return the subscript value or the allocated
value of the outer object, whichever is larger.