A more aggressive approach
Always punt to outermost object for flex arrays.  We don't need to
traverse upwards and all that, if we have a flex array reference, use
it.

This still doesn't work for, e.g.:

struct A
{
  char pad;
  struct
  {
    char pad;
    char data[];
  } s;
} a;

with __bos (&a.s, 1);

which should give the allocated amount.  This is because the flex array
function doesn't detect it as a flex array.  One simple thing could be
to treat last elements specially at all times, ignoring the flex array
setting.