tree-object-size: Fix interaction of overestimates with negative offsets
When the object size estimate is returned for a PHI node, it is the
maximum possible value, which is fine in isolation.  When combined with
negative offsets however, it may sometimes end up in zero size because
the resultant size was larger than the wholesize, leading
size_for_offset to conclude that there's a potential underflow.  Fix
this by allowing a non-strict mode to size_for_offset, which
conservatively returns the size (or wholesize) in case of a negative
offset.

The *correct* fix for this would be to maintain the full set of size
options from PHI or COND_EXPR at all times (like we end up doing for
dynamic sizes) and return the best available size in the context of the
expression it appears in, but that would need static object sizes to be
evaluated from the dynamic expressions, which could be a whole another
project.

Signed-off-by: Siddhesh Poyarekar <siddhesh@gotplt.org>