WIP: Minimum size computation should bail out on allocation failure
When the allocator call is visible, its argument(s) are directly passed on as
object size for maximum as well as minimum sizes.  For ill formed programs that
don't check for allocation failure, this could result in the object size being
incorrectly reported as a positive number, which is problematic for minimum
size.

Fix this by folding in a NULL check into the object size expression, having it
bail out with a 0 value for minimum size when the pointer is NULL.

TODO: Fix up tests and run a bootstrap.